backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
MetaTrader5==5.0.5370
python-dotenv==1.0.0
asyncpg==0.29.0
aioredis==2.0.1
redis==5.0.1
bcrypt==4.1.2
cryptography==42.0.0

# Scheduler
croniter==2.0.1
numpy==1.26.2
